% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_point_model}
\alias{change_point_model}
\title{Generate samples from a change point model}
\usage{
change_point_model(n, mechanism = contaminated_sample_t, cpt = NA, kappa = 1)
}
\arguments{
\item{n}{The number of samples to generate. A positive integer.}

\item{mechanism}{A function for the sample generating mechanism with two inputs, n and epsilon.
Defaults to \code{contaminated_sample_t}.}

\item{cpt}{Location of changepoint. A positive integer less than n or \code{NA}, which
means no change point. Defaults to NA}

\item{kappa}{Change in mean at the change point.}
}
\value{
Samples from a t-distribution contaminated with normal distributed data.
}
\description{
Generate samples from a change point model with user-specified
contamination level and underlying data generating mechanism.
}
