% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{contaminated_sample_t}
\alias{contaminated_sample_t}
\title{t-distribution contaminated with normal samples}
\usage{
contaminated_sample_t(n, df = 3, epsilon = 0, mu = 0)
}
\arguments{
\item{n}{The number of samples to generate. A positive integer.}

\item{df}{The degrees of freedom of the t-distribution. A positive number, defaults to 3.}

\item{epsilon}{Contamination level for the Normal(0,25) distribution. A number in (0,1)}

\item{mu}{Mean of the inlier distribution.}
}
\value{
Samples from a t-distribution contaminated with normal
}
\description{
Generate samples from a t-distribution contaminated with normal
distribution with large variance
}
