% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{rlaplace_hd}
\alias{rlaplace_hd}
\title{High dimensional Laplace distribution}
\usage{
rlaplace_hd(n, p = 1, s = 1, epsilon = 0, mu = rep(0, p))
}
\arguments{
\item{n}{The number of samples to generate. A positive integer.}

\item{p}{Dimension}

\item{s}{Scale parameter of the Laplace distribution. Defaults to 1.}

\item{epsilon}{Contamination level for the Normal(0,?) distribution. A number in (0,1).}

\item{mu}{Mean of the inlier distribution.}
}
\value{
Samples from a t-distribution contaminated with normal
}
\description{
Generate samples from a t-distribution contaminated with normal
distribution
}
