% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{RobustMeanTest}
\alias{RobustMeanTest}
\title{Robust Mean Testing routine}
\usage{
RobustMeanTest(Y, kappa0, delta, epsilon, C_gamma = 1)
}
\arguments{
\item{Y}{dataset as matrix of n rows and p columns}

\item{kappa0}{Norm of mean difference to be detected}

\item{delta}{confidence level}

\item{epsilon}{contamination level}

\item{C_gamma}{tuning parameter}
}
\description{
Canconne et al (2023)
}
