% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{rowsum_filter}
\alias{rowsum_filter}
\title{Rowsum_filter subroutine}
\usage{
rowsum_filter(Y, w, u)
}
\arguments{
\item{Y}{dataset as matrix of n rows and p columns}

\item{w}{Processed weights}

\item{u}{contamination level}
}
\description{
Canconne et al (2023)
}
