% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{rt_hd}
\alias{rt_hd}
\title{High dimensional t-distribution contaminated with normal samples}
\usage{
rt_hd(n, p = 1, df = 3, epsilon = 0, mu = rep(0, p))
}
\arguments{
\item{n}{The number of samples to generate. A positive integer.}

\item{p}{Dimension}

\item{df}{The degrees of freedom of the t-distribution. A positive number, defaults to 3.}

\item{epsilon}{Contamination level for the Normal(0,?) distribution. A number in (0,1).}

\item{mu}{Mean of the inlier distribution.}
}
\value{
Samples from a t-distribution contaminated with normal
}
\description{
Generate samples from a t-distribution contaminated with normal
distribution
}
