% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rume}
\alias{rume}
\title{RUME estimator}
\usage{
rume(X, epsilon = 0, delta = 0.05)
}
\arguments{
\item{X}{Univariate dataset}

\item{epsilon}{Contamination level, a number in (0,1) 0}

\item{delta}{Confidence level, a number in (0,1)}
}
\value{
The robust estimate of mean
}
\description{
Robust Univariate Mean Estimation, suggested by Prasad et al. (2021)
}
