% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rumedian}
\alias{rumedian}
\title{rumedian changepoint algorithm}
\usage{
rumedian(online_data, epsilon = 0, alpha = 0.1)
}
\arguments{
\item{online_data}{The complete dataset to analyse. A vector of numbers.}

\item{epsilon}{Contamination level. A number in (0,1).}

\item{alpha}{Type I error. A number in (0,1).}
}
\value{
Location where change point is detected
}
\description{
Runs rumedian algorithm on \code{online_data}.
}
