% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rumedian_v}
\alias{rumedian_v}
\title{rumedian changepoint algorithm}
\usage{
rumedian_v(
  online_data,
  sigma,
  v = 2,
  epsilon = 0,
  alpha = 0.1,
  C1 = 1.59,
  C2 = 2.25
)
}
\arguments{
\item{online_data}{The complete dataset to analyse. A vector of numbers.}

\item{epsilon}{Contamination level. A number in (0,1).}

\item{alpha}{Type I error. A number in (0,1).}
}
\value{
Location where change point is detected
}
\description{
Runs the rumedian algorithm on \code{online_data}.
}
