% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional.R
\name{spectral_filter}
\alias{spectral_filter}
\title{Spectral_filter subroutine}
\usage{
spectral_filter(Y, gamma2)
}
\arguments{
\item{Y}{dataset as matrix of n rows and p columns}

\item{gamma2}{Tuning parameter}
}
\description{
Canconne et al (2023)
}
