% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_point_model}
\alias{change_point_model}
\title{Generate samples from a change point model}
\usage{
change_point_model(
  n,
  mechanism = contaminated_sample_t,
  epsilon = 0,
  cpt = NA,
  kappa = 1
)
}
\arguments{
\item{n}{The number of samples to generate. A positive integer.}

\item{epsilon}{Contamination level for the Normal(0,25) distribution. A number in (0,1)}

\item{df}{The degrees of freedom of the t-distribution. A positive number, defaults to 3.}

\item{mu}{Mean of the inlier distribution.}
}
\value{
Samples from a t-distribution contaminated with normal
}
\description{
Generate samples from a change point model with user-specified
contamination level and underlying data generating mechanism.
}
